chain applied in statemaker included the pivot block when it should not have
This broke the ability of state to sync between the block hypothesising stage (mining) and whne it was imported. THe issue was detected because the tx_tracker dapp was recording an incorrect height which threw the state hash out. The incorrect height implied the wrong block was being submitted to _on_transaction.